ja manda imagem mas ainda nao corrigi o alarme. nao funcionava por 2 motivos: 1. comparacao de char c/ unsigned char. 2. nos filesize fazias cast para unsigned char e este extendia o sinal (ex:10000000 -> 1111111110000000)

+++ lab1/prat5/FileFuncs.c
@@ -41,7 +41,7 @@ long getFileBytes(char* filename, char** dest_buf)
 }
 
 
-int save2File(char* data, long data_size, const char* filename)
+int save2File(char* data, int data_size, const char* filename)
 {
 	FILE *pFile;
